refactor: moved predict feature dialogs to dialog folder
