+++ lib/dialogs/predict_mint/bloc/mint_button_state.dart
