+++ lib/dialogs/predict_no/bloc/no_button_state.dart
