+++ lib/dialogs/predict_redeem/bloc/redeem_button_event.dart
