+++ lib/dialogs/predict_yes/bloc/yes_button_event.dart
