+++ lib/prediction/widgets/prompt_details.dart
@@ -1,5 +1,6 @@
+import 'package:ax_dapp/dialogs/predict_no/no_button.dart';
+import 'package:ax_dapp/dialogs/predict_yes/yes_button.dart';
 import 'package:ax_dapp/predict/models/prediction_model.dart';
-import 'package:ax_dapp/prediction/widgets/buttons/buttons.dart';
 import 'package:ax_dapp/service/custom_styles.dart';
 import 'package:ax_dapp/util/colors.dart';
 import 'package:flutter/material.dart';
